use a smallvec inside `IntersectionType` itself
